Revert mypy changes (#1836)
* Revert mypy changes

* Remove unused import

+++ torchgeo/datasets/benin_cashews.py
@@ -384,8 +384,7 @@ class BeninSmallHolderCashews(NonGeoDataset):
             dtype=np.uint8,
         )
 
-        mask = torch.from_numpy(mask_data)
-        mask = mask.long()
+        mask = torch.from_numpy(mask_data).long()
         return mask
 
     def _check_integrity(self) -> bool:
